requirements.txt
asgiref==3.9.1
boto3==1.40.59
botocore==1.40.59
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.6
django-cleanup==9.0.0
django-crontab==0.7.1
django-ses==4.4.0
djangorestframework==3.16.1
fire==0.7.1
idna==3.10
jmespath==1.0.1
markdown-it-py==4.0.0
mdurl==0.1.2
pillow==11.3.0
Pygments==2.19.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
rav==0.1.1
requests==2.32.5
rich==14.1.0
ruff==0.14.3
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
termcolor==3.1.0
urllib3==2.5.0
